Removed unnecessary code and little refactor
